fusion du fichier Heros.java de Ugo et de P.;le main des Heros.java est déplacé dans le main de LaFenetre.j (mis en commentaires);petites corrections des erreurs liés aux fichiers Heros.j et PanelEdition.j
